+++ apps/vmq_server/src/vmq_server_sup.erl
@@ -44,6 +44,7 @@ start_link() ->
 init([]) ->
     maybe_change_nodename(),
     persistent_term:put(subscribe_trie_ready, 0),
+    init_systemd_notify(),
     {ok,
         {{one_for_one, 5, 10}, [
             ?CHILD(vmq_config, worker, []),
@@ -72,4 +73,14 @@ maybe_change_nodename() ->
             %% we ignore if the node has the same name
             %% or if more than one node is returned (clustered)
             ignore
+    end.
+
+init_systemd_notify() ->
+    %% systemd -> set status to ready and start watchdog
+    Pid = os:getpid(),
+    systemd:notify(ready),
+    case os:getenv("WATCHDOG_PID") of
+        false -> systemd:watchdog(enable);
+        Pid -> systemd:watchdog(enable);
+        _ -> false
     end.